% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawExpression.R
\name{drawExpression}
\alias{drawExpression}
\title{Generate a three dimensional heatmap with x axis representing the
developmental stages, y axis representing the FlyBase Gene IDs, and z (the
color of each unit cell) representing the level of expression(intensity).
The expression for each gene is composed of two different types of expression
data. One is specific to embryogenesis expression where the data contain gene
expression at 1hr, 2hr, 3hr, 4hr, 5hr, 6hr, 8hr, 10hr, 12hr, 14hr, 16hr,
18hr, and 20hr. The other is specific to the life cycle where the data
contain gene expression spanning the whole life cycle of the fly, including
embryo 0-2hr(E0-2hr), embryo 4-6hr(E4-6hr), embryo 10-12hr(E10-12hr),
embryo 18-20hr(E18-20hr), 1st instar larva(L1), 2nd instar larva(L2),
3rd instar larva(L3), late 3rd instar larva(late L3), white prepupa(WPP),
pupal stage 2(P2), pupal stage 3(P3), pupal stage 4(P4), pupal stage 5(P5),
Adult young female(Ayf), Adult young male(Aym), Adult female(Af), and
Adult male(Am). The user can choose which heatmap to draw by changing the
type parameter("Embryogenesis" or "LifeCycle").}
\usage{
drawExpression(expressionData, typeGraph, graphTitle)
}
\arguments{
\item{expressionData}{the list of expressionData retrieved from FlyBase API
using getAllExpression() function.}

\item{typeGraph}{The type of expression data the heat map contains. User can
choose from "Embryogenesis" or "LifeCycle".}

\item{graphTitle}{A string used as the title of the heatmap generated}
}
\description{
Generate a three dimensional heatmap with x axis representing the
developmental stages, y axis representing the FlyBase Gene IDs, and z (the
color of each unit cell) representing the level of expression(intensity).
The expression for each gene is composed of two different types of expression
data. One is specific to embryogenesis expression where the data contain gene
expression at 1hr, 2hr, 3hr, 4hr, 5hr, 6hr, 8hr, 10hr, 12hr, 14hr, 16hr,
18hr, and 20hr. The other is specific to the life cycle where the data
contain gene expression spanning the whole life cycle of the fly, including
embryo 0-2hr(E0-2hr), embryo 4-6hr(E4-6hr), embryo 10-12hr(E10-12hr),
embryo 18-20hr(E18-20hr), 1st instar larva(L1), 2nd instar larva(L2),
3rd instar larva(L3), late 3rd instar larva(late L3), white prepupa(WPP),
pupal stage 2(P2), pupal stage 3(P3), pupal stage 4(P4), pupal stage 5(P5),
Adult young female(Ayf), Adult young male(Aym), Adult female(Af), and
Adult male(Am). The user can choose which heatmap to draw by changing the
type parameter("Embryogenesis" or "LifeCycle").
}
\examples{
\dontrun{
expressionData <- getAllExpression(geneList = geneListExample)

# For generating a heatmap for gene expression specific to embryogenesis.
drawExpression(expressionData = expressionData,
               typeGraph = "Embryogenesis",
               graphTitle = "Embryogenesis Gene Expression")

# For generating a heatmap for gene expression specific to life cycle,
# including 4 embryo stages, 4 larva stages, 1 prepupal stage, 4 pupal
# stages, and 2 adult stages each for male and female flies.
drawExpression(expressionData = expressionData,
               typeGraph = "LifeCycle",
               graphTitle = "Life Cycle Gene Expression")
}

}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
