% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expressionDataExample}
\alias{expressionDataExample}
\title{Expression data example generated by getAllExpression() function saved in
an rda file.}
\format{
dataframe
}
\usage{
expressionDataExample
}
\description{
Expression data example generated by getAllExpression() function saved in
an rda file.
}
\examples{
\dontrun{
 expressionDataExample
}

}
\keyword{datasets}
