% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneListExample}
\alias{geneListExample}
\title{Random List of FlyBase Gene IDs used to demonstrate the usage of the package.}
\format{
list
}
\usage{
geneListExample
}
\description{
Random List of FlyBase Gene IDs used to demonstrate the usage of the package.
}
\examples{
\dontrun{
 geneListExample
}

}
\keyword{datasets}
