% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndParseData.R
\name{getAllSepuences}
\alias{getAllSepuences}
\title{Generate a text file containing all sequences for the genes ids input}
\usage{
getAllSepuences(geneFile, seqType)
}
\arguments{
\item{geneFile}{the list of genes for sequence retrieval as a txt file,
each gene on a newline.}

\item{seqType}{the tyoe if sequence you want to retrieve, any one choice
from SEQUENCETYPE}
}
\description{
Generate a text file containing all sequences for the genes ids input
}
\examples{
#Example 1:
#This will retrieve the sequence data, then export an txt file
with the list of genes and their sequences
An example input file is called geneList.txt in tests folder.

> FlyBaseInR::getAllSequences(tests/geneList.txt, "exon")
> Please check output file "exonSeqeunces"in in your current
 working directory.

}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
