% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllExpression.R
\name{getExpression}
\alias{getExpression}
\title{INTERNAL FUNCTION: Get expression data from proteomic label-free
quantification (LFQ) expression values from FlyBase for a single gene using
flyID (FlyBase GeneID).}
\usage{
getExpression(flyID)
}
\arguments{
\item{flyID}{A FlyBase ID that FlyBase uses to identify gene}
}
\value{
A list of expression LFQ intensity and corresponding developmental
stages of the gene(by flyID).
}
\description{
This function Will use FlyBase internal API function to retrieve feature
using specified FlyBase gene ID.
}
\references{
Wickham H (2022). \emph{httr: Tools for Working with URLs and HTTP}.
R package version 1.4.4, \href{https://CRAN.R-project.org/package=httr}{link}
}
\author{
{Huilin Niu, \email{ huilin.niu@mail.utoronto.ca}}
}
