% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndParseData.R
\name{getExpressiondf}
\alias{getExpressiondf}
\title{INTERNAL FUNCTION: read a text file and return in a list.}
\usage{
getExpressiondf(geneList)
}
\arguments{
\item{geneList}{A set of genes for analysis}
}
\value{
A dataframe containing all information for writing the
expression table
}
\description{
This function Will use FlyBase internal API function to retrieve feature
using specified FlyBase gene ID.
}
\author{
{Huilin Niu, \email{ huilin.niu@mail.utoronto.ca}}
}
