% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSequences.R
\name{getSequence}
\alias{getSequence}
\title{INTERNAL FUNCTION: Get one sequence from FlyBase API using FlyBase ID and
the sequence type specified in the SEQUENCETYPE global variable.}
\usage{
getSequence(seqType, flyID)
}
\arguments{
\item{seqType}{The sequence type for retrieval specified by a string. The
available types include: FBgn, FBtr, FBpp, FBti, FBtp, exon,
intron, five_prime_utr, three_prime_utr, CDS. The sequences type that the
abbreviation represents are listed below.
\enumerate{
\item FBgn: The sequence of the gene span.
\item FBtr: The sequence of transcripts that are typed as messenger RNAs (mRNA).
\item FBpp: The resulting protein sequence from protein coding transcripts.
\item FBtp: Transgenetic construct sequence.
\item exon: Exon sequence split into each exon.
\item intron: Translated intron sequence split into each intron.
\item five_prime_utr: The sequence of 5' untranslated regions.
\item three_prime_utr: The sequence of 3' untranslated regions.
\item CDS: The contiguous protein coding sequence which begins with, and
includes, a start codon and ends with, and includes, a stop codon.
}}

\item{flyID}{A FlyBase ID that FlyBase uses to identify gene, start with
FBgn.}
}
\value{
A list of sequence/sequences
}
\description{
INTERNAL FUNCTION: Get one sequence from FlyBase API using FlyBase ID and
the sequence type specified in the SEQUENCETYPE global variable.
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
