% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndParseData.R
\name{parseFile}
\alias{parseFile}
\title{INTERNAL FUNCTION: read a text file and return in a list.}
\usage{
parseFile(geneFile)
}
\arguments{
\item{geneFile}{a text file with genes seperated by new line, the input
should be the file path.}
}
\value{
A list of genes
}
\description{
This function Will use FlyBase internal API function to retrieve feature
using specified FlyBase gene ID.
}
\author{
{Huilin Niu, \email{ huilin.niu@mail.utoronto.ca}}
}
