% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndParseData.R
\name{printFinish}
\alias{printFinish}
\title{INTERNAL FUNCTION: Print finish message for users}
\usage{
printFinish(outputFileName)
}
\arguments{
\item{outputFileName}{the name of the finihsed output file}
}
\description{
INTERNAL FUNCTION: Print finish message for users
}
\author{
{Huilin Niu, \email{ huilin.niu@mail.utoronto.ca}}
}
