% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndParseData.R
\name{printWarning}
\alias{printWarning}
\title{INTERNAL FUNCTION: Request response warnings}
\usage{
printWarning(status, flyID)
}
\arguments{
\item{status}{The status code the request returned}

\item{flyID}{A FlyBase ID that FlyBase uses to identify gene}
}
\description{
This function will take different status code of the request, and print
corresponding warning messages to users
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
