% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFlyBaseInR.R
\name{runFlyBaseInR}
\alias{runFlyBaseInR}
\title{Launch Shiny App for FlyBaseInR}
\usage{
runFlyBaseInR()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for FlyBaseInR.
The shiny app allows user to analyze and visualize expression levels of
specific genes in FlyBase Drosophila Melanogaster database via an
interactive platform.
}
\details{
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
FlyBaseInR::runFlyBaseInR()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
