% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawExpression.R
\name{sortHeatMapData}
\alias{sortHeatMapData}
\title{INTERNAL FUNCTION: Extract Embryogenesis or LifeCycle expression data from
expressionData gathered from a list of genes.}
\usage{
sortHeatMapData(expressionData, typeData)
}
\arguments{
\item{expressionData}{the list of gene expression including embryogenesis
and life cycle stages.}

\item{typeData}{The type of expression data the heat map contains. User can
choose from "Embryogenesis" or "LifeCycle".}
}
\value{
A data frame containing only Embryogenesis or LifeCycle expression,
specified by the user.
}
\description{
INTERNAL FUNCTION: Extract Embryogenesis or LifeCycle expression data from
expressionData gathered from a list of genes.
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
