% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawExpression.R
\name{drawExpression}
\alias{drawExpression}
\title{Generate a heatmap containing all expression intensity level for the genes
ids input}
\usage{
drawExpression(geneFile, stage)
}
\arguments{
\item{geneFile}{the list of genes for sequence retrieval as a txt file,
each gene on a newline.}

\item{stage}{choose from embryogenesis and life cycle}
}
\description{
Generate a heatmap containing all expression intensity level for the genes
ids input
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
