% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndParseData.R
\name{getAllExpression}
\alias{getAllExpression}
\title{Generate a text file containing all sequences for the genes ids input}
\usage{
getAllExpression(geneFile)
}
\arguments{
\item{geneFile}{the list of genes for sequence retrieval as a txt file,
each gene on a newline.}
}
\description{
Generate a text file containing all sequences for the genes ids input
}
\examples{
#Example 1:
#This will retrieve the expression data, then export an txt file
with the list of genes and their expression
An example input file is called geneList.txt in tests folder.

> FlyBaseInR::getAllExpression("tests/geneList.txt")
> Please check output file expression.txt in your current working directory.

}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
