% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllExpression.R
\name{getAllExpression}
\alias{getAllExpression}
\title{Generate a text file containing all sequences for the genes ids input to
keep record, and return the expression data as a data frame.}
\usage{
getAllExpression(geneList)
}
\arguments{
\item{geneList}{A set of genes represented by FlyBase gene ID for retrival
of expression.}
}
\value{
a list containing the expression data of all available
expressions.Each row contains the expression data from embryogenesis to adult
stages, and the name of the gene is specified in the first column of each
row.
}
\description{
Generate a text file containing all sequences for the genes ids input to
keep record, and return the expression data as a data frame.
}
\examples{
\dontrun{
getAllExpression(geneListExample)
}

}
\references{
R Core Team (2022). R: A language and environment for statistical computing.
\emph{R Foundation for Statistical Computing} , Vienna, Austria.
\href{https://www.R-project.org/}{link}
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
