% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSequences.R
\name{getAllSequences}
\alias{getAllSequences}
\title{Generate a text file containing all sequences for the FlyBase ids from user
input.}
\usage{
getAllSequences(geneList, seqType)
}
\arguments{
\item{geneList}{the list of genes for sequence retrieval.}

\item{seqType}{the type if sequence you want to retrieve, any one choice
from SEQUENCETYPE}
}
\description{
Generate a text file containing all sequences for the FlyBase ids from user
input.
}
\examples{
\dontrun{
getAllSequences(geneList = geneListExample, seqType = "FBpp")
}

}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
