% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndParseData.R
\name{getExpression}
\alias{getExpression}
\title{INTERNAL FUNCTION: Get expression data from proteomic label-free
quantification (LFQ) expression values from FlyBase using
flyID (FlyBase GeneID)}
\usage{
getExpression(flyID)
}
\arguments{
\item{flyID}{A FlyBase ID that FlyBase uses to identify gene}
}
\value{
A list of expression LFQ intensity and corresponsing developmental
stages of the gene(by flyID).
}
\description{
This function Will use FlyBase internal API function to retrieve feature
using specified FlyBase gene ID.
}
\author{
{Huilin Niu, \email{ huilin.niu@mail.utoronto.ca}}
}
