% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllExpression.R
\name{getExpressiondf}
\alias{getExpressiondf}
\title{INTERNAL FUNCTION: Get a expression data for a list of genes
for expression analysis.}
\usage{
getExpressiondf(geneList)
}
\arguments{
\item{geneList}{A set of genes for analysis}
}
\value{
A dataframe containing all information for writing the
expression table
}
\description{
INTERNAL FUNCTION: Get a expression data for a list of genes
for expression analysis.
}
\references{
Wickham H, François R, Henry L, Müller K (2022).
\emph{dplyr: A Grammar of Data Manipulation}. R package version 1.0.10,
\href{https://CRAN.R-project.org/package=dplyr}{link}
}
\author{
{Huilin Niu, \email{ huilin.niu@mail.utoronto.ca}}
}
