% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndParseData.R
\name{getSequence}
\alias{getSequence}
\title{INTERNAL FUNCTION: Get seqeunce from FlyBase using given feature and flyID}
\usage{
getSequence(seqType, flyID)
}
\arguments{
\item{seqType}{The type of sequence which we want to retrieve from FlyBase,
including nuceotide sequence, mRNA sequences, and amino acid sequences.The
available subtypes include :  FBgn, FBtr, FBpp, FBcl, FBsf, FBti, FBtp, exon,
intron, five_prime_utr, three_prime_utr, CDS}

\item{flyID}{A FlyBase ID that FlyBase uses to identify gene}
}
\value{
A list of sequences that belong to the flyID and the sequence type
}
\description{
This function Will use FlyBase internal API function to retrieve feature
using specified FlyBase gene ID.
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
