% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSequences.R
\name{parseFile}
\alias{parseFile}
\title{Read a text file and return in a list. This function can be used to format
FlyBase ids into a list}
\usage{
parseFile(geneFile)
}
\arguments{
\item{geneFile}{a text file with genes separated by new line, the input
should be the name of the file, and the file should be stored
in the current working directory for the user.}
}
\value{
A list of genes
}
\description{
This function will take the user input text file, and format the input
FlyBase IDs to list. The output can be used in getAllSequences() or
getAllExpression() function.
}
\author{
{Huilin Niu, \email{ huilin.niu@mail.utoronto.ca}}
}
