% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFlyBaseInR.R
\name{runFlyBaseInR}
\alias{runFlyBaseInR}
\title{Launch Shiny App for FlyBaseInR}
\usage{
runFlyBaseInR()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for FlyBaseInR.
The shiny app allows user to analyze and visualize expression levels of
specific genes in FlyBase \emph{Drosophila Melanogaster} database via an
interactive platform. The code has been placed in
\code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
FlyBaseInR::runFlyBaseInR()
}

}
\references{
Chang W, Cheng J, Allaire J, Sievert C, Schloerke B, Xie Y, Allen J,
McPherson J, Dipert A, Borges B (2022). \emph{shiny: Web Application Framework
for R}. R package version 1.7.3.9001,
\href{https://shiny.rstudio.com/}{link}
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
