% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawExpression.R
\name{sortHeatMapData}
\alias{sortHeatMapData}
\title{INTERNAL FUNCTION: Sort the data into embryogenesis data or life cycle data.}
\usage{
sortHeatMapData(geneFile, stage)
}
\arguments{
\item{geneFile}{the list of genes for sequence retrieval as a txt file,
each gene on a newline.}

\item{stage}{Stage specifying which part of the list is used for extraction}
}
\value{
Two dataframe with embryogenesis pattern and another with
life cycle pattern
}
\description{
INTERNAL FUNCTION: Sort the data into embryogenesis data or life cycle data.
}
\author{
{Huilin Niu, \email{huilin.niu@mail.utoronto.ca}}
}
